Refactor: [백엔드] 전적 정보(Search-service) 에서 소환사 정보들만 추출하여 소환사 DB 저장
[백엔드] 전적 정보 조회 기능 -> 소환사 정보들만 추출하여 소환사 DB 저장

+++ Search-service/src/main/java/com/ggwp/searchservice/summoner/dto/RequestCreateSummonerDto.java
@@ -12,11 +12,10 @@ import java.time.ZoneOffset;
 @NoArgsConstructor
 @Builder
 public class RequestCreateSummonerDto {
+
     private String id;
 
-    private Long profileIconId;
-
-//    private String accountId;
+    private int profileIconId;
 
     private String puuid;
 
@@ -25,6 +24,8 @@ public class RequestCreateSummonerDto {
     private Long revisionDate;
 
     private Long summonerLevel;
+
+    //    private String accountId;
 
     // DTO를 -> Entity 로 변경
     public Summoner toEntity() {
