+++ Search-service/src/main/java/com/ggwp/searchservice/summoner/dto/ResponseGetSummonerDto.java
@@ -16,7 +16,7 @@ public class ResponseGetSummonerDto {
 
     private String id;
 
-    private Long profileIconId;
+    private int profileIconId;
 
 //    private String accountId;
